requirements.txt
alembic==1.15.2
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
boto3==1.38.8
botocore==1.38.8
build==1.2.2.post1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-rest-framework==0.1.0
django-rosetta==0.10.2
django-ses==4.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
git-filter-repo==2.47.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.0
oauthlib==3.2.2
packaging==25.0
payme-pkg==3.0.24
pillow==11.2.1
pip-tools==7.4.1
polib==1.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
resend==2.9.0
rpds-py==0.25.1
s3transfer==0.12.0
service-identity==24.2.0
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
text-unidecode==1.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
whitenoise==6.9.0
zope.interface==7.2
